requirements.txt
# =============================================================================
# GatheRing Framework - Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# For development: pip install -r requirements-dev.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies (Required)
# -----------------------------------------------------------------------------

# Configuration and validation
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
email-validator>=2.0,<3.0

# -----------------------------------------------------------------------------
# Database (PostgreSQL)
# -----------------------------------------------------------------------------

# SQLAlchemy ORM
sqlalchemy>=2.0,<3.0

# PostgreSQL async driver
asyncpg>=0.29.0,<1.0

# PostgreSQL sync driver (for migrations)
psycopg2-binary>=2.9,<3.0

# Vector embeddings support
pgvector>=0.3.0,<1.0

# Database migrations
alembic>=1.13,<2.0

# -----------------------------------------------------------------------------
# LLM Providers
# -----------------------------------------------------------------------------

# OpenAI (updated to latest stable)
openai>=1.30,<2.0

# Anthropic (updated to latest stable)
anthropic>=0.35,<1.0

# Ollama (Local LLM - updated)
ollama>=0.3,<1.0

# LangChain ecosystem (optional, for advanced features)
# langchain>=0.1.0
# langgraph>=0.0.20
# langchain-openai>=0.0.5
# langchain-anthropic>=0.0.1
# langchain-community>=0.0.10

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------

# JWT tokens
python-jose[cryptography]>=3.3,<4.0

# Password hashing
passlib[bcrypt]>=1.7,<2.0

# Bcrypt backend
bcrypt>=4.0,<5.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------

# CLI and display
click>=8.1,<9.0
rich>=13.0,<14.0

# HTTP client
httpx>=0.25,<1.0

# Async support
aiofiles>=23.0,<25.0

# Scheduling
croniter>=2.0,<3.0

# Redis cache
redis>=5.0,<6.0

# -----------------------------------------------------------------------------
# Observability (OpenTelemetry)
# -----------------------------------------------------------------------------

# Core OpenTelemetry
opentelemetry-api>=1.20,<2.0
opentelemetry-sdk>=1.20,<2.0

# OTLP exporters (Jaeger, Prometheus)
opentelemetry-exporter-otlp-proto-grpc>=1.20,<2.0

# Auto-instrumentation
opentelemetry-instrumentation-requests>=0.41,<1.0
opentelemetry-instrumentation-httpx>=0.41,<1.0

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------

# FastAPI
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
python-multipart>=0.0.6,<1.0
